### 2024-11-18 #7499 @editorialbot
[PRE REVIEW]: foamlib: A modern Python package for working with OpenFOAM
**Submitting author:** @gerlero (<a href="http://orcid.org/0000-0002-5138-0328">Gabriel S. Gerlero</a>)
**Repository:** https://github.com/gerlero/foamlib
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.7.0
