### 2024-11-18 #7499 @kyleniemeyer
Hello @gerlero, thanks for your submission to JOSS. We're currently managing a large backlog of submissions and the editor most appropriate for your area is already rather busy.

While we are waiting, could you please some potential reviewers from our database? (Please do not actually tag them with @ though)
