### 2024-11-18 #7499 @gerlero
> While we are waiting, could you please some potential reviewers from our database? (Please do not actually tag them with @ though)

**Possible reviewers** in database:
- [MakisH](https://reviewers.joss.theoj.org/reviewers/3047)
- [chennachaos](https://reviewers.joss.theoj.org/reviewers/605)

**More possible reviewers** (JOSS paper authors not in database):
- philipcardiff
- blebon

> For now, we will need to waitlist this paper and process it as the queue reduces. Thanks for your patience!

Thank you!
