### 2024-11-19 fa680f9
Improve ASCII field parsing performance
