### 2024-11-23 b8b9c2c
Update pyproject.toml
