### 2024-11-23 48ba18e
Update type hint
