### 2024-11-25 3e12602
Update parsing of lists of keyword entries
