+++ tests/test_files/test_parsing.py
@@ -63,16 +63,16 @@ def test_parse_value() -> None:
         "simpleGrading",
         [1, 1, 1],
     )
-    assert Parsed(b"(a b; c d;)")[()] == [{"a": "b"}, {"c": "d"}]
+    assert Parsed(b"(a b; c d;)")[()] == [("a", "b"), ("c", "d")]
     assert Parsed(b"(a {b c;} d {e g;})")[()] == [
-        {"a": {"b": "c"}},
-        {"d": {"e": "g"}},
+        ("a", {"b": "c"}),
+        ("d", {"e": "g"}),
     ]
-    assert Parsed(b"(a (0 1 2); b {})")[()] == [{"a": [0, 1, 2]}, {"b": {}}]
+    assert Parsed(b"(a (0 1 2); b {})")[()] == [("a", [0, 1, 2]), ("b", {})]
     assert Parsed(b"(water oil mercury air)")[()] == ["water", "oil", "mercury", "air"]
     assert Parsed(b"div(phi,U)")[()] == "div(phi,U)"
     assert Parsed(b"((air and water) { type constant; sigma 0.07; })")[()] == [
-        {"(air and water)": {"type": "constant", "sigma": 0.07}}
+        (["air", "and", "water"], {"type": "constant", "sigma": 0.07})
     ]
     assert Parsed(b"[]")[()] == FoamFile.DimensionSet()
     assert Parsed(b"object f.1;")[("object",)] == "f.1"
