### 2024-11-26 787fea6
Refactor parsing
