### 2024-11-27 f0e7b13
Update pytest configuration
