+++ pyproject.toml
@@ -76,9 +76,6 @@ version = {attr = "foamlib.__version__"}
 [tool.setuptools.package-data]
 "foamlib" = ["py.typed"]
 
-[tool.pytest.ini_options]
-asyncio_default_fixture_loop_scope = "function"
-
 [tool.mypy]
 packages = [
     "foamlib",
