+++ pyproject.toml
@@ -38,7 +38,7 @@ dynamic = ["version"]
 
 [project.optional-dependencies]
 numpy = [
-    "numpy>=1.25.0,<3; python_version>='3.10'",
+    "numpy>=1,<3; python_version>='3.10'",
     "numpy>=1,<3"
 ]
 lint = ["ruff"]
