### 2024-12-04 60119f9
Update type hints
