### 2024-12-04 31daf48
Handle lists of dictionaries without keywords
