### 2024-12-04 4215fdc
Bump version to 0.8.0
