+++ pyproject.toml
@@ -1,9 +1,10 @@
 [build-system]
-requires = ["hatchling"]
-build-backend = "hatchling.build"
+requires = ["uv>=0.5,<0.6"]
+build-backend = "uv"
 
 [project]
 name = "foamlib"
+version = "0.8.0"
 description = "A Python interface for interacting with OpenFOAM"
 readme = "README.md"
 requires-python = ">=3.7"
@@ -33,8 +34,6 @@ dependencies = [
     "pyparsing>=3,<4",
     "typing-extensions>=4,<5; python_version<'3.11'",
 ]
-
-dynamic = ["version"]
 
 [project.optional-dependencies]
 numpy = [
@@ -69,9 +68,6 @@ dev = [
 Homepage = "https://github.com/gerlero/foamlib"
 Repository = "https://github.com/gerlero/foamlib"
 Documentation = "https://foamlib.readthedocs.io"
-
-[tool.hatch.version]
-path = "foamlib/__init__.py"
 
 [tool.mypy]
 packages = [
