### 2024-12-05 dbcb83c
Update serialization
