### 2024-12-05 ac8399e
Integrate NumPy
