+++ foamlib/_files/_serialization.py
@@ -1,34 +1,25 @@
 from __future__ import annotations
 
-import array
-import itertools
 import sys
 from enum import Enum, auto
-from typing import cast, overload
+from typing import overload
 
 if sys.version_info >= (3, 9):
-    from collections.abc import Mapping, Sequence
+    from collections.abc import Mapping
 else:
-    from typing import Mapping, Sequence
+    from typing import Mapping
+
+import numpy as np
 
 from ._parsing import parse_data
-from ._types import Data, Dimensioned, DimensionSet, Entry
-from ._util import is_sequence
-
-try:
-    import numpy as np
-
-    numpy = True
-except ModuleNotFoundError:
-    numpy = False
+from ._types import Data, Dimensioned, DimensionSet, Entry, is_sequence
 
 
 class Kind(Enum):
     DEFAULT = auto()
     SINGLE_ENTRY = auto()
     ASCII_FIELD = auto()
-    DOUBLE_PRECISION_BINARY_FIELD = auto()
-    SINGLE_PRECISION_BINARY_FIELD = auto()
+    BINARY_FIELD = auto()
     DIMENSIONS = auto()
 
 
@@ -41,9 +32,29 @@ def normalize(data: Entry, *, kind: Kind = Kind.DEFAULT) -> Entry: ...
 
 
 def normalize(data: Entry, *, kind: Kind = Kind.DEFAULT) -> Entry:
-    if numpy and isinstance(data, np.ndarray):
+    if kind in (Kind.ASCII_FIELD, Kind.BINARY_FIELD):
+        if is_sequence(data):
+            try:
+                arr = np.asarray(data)
+            except ValueError:
+                pass
+            else:
+                if not np.issubdtype(arr.dtype, np.floating):
+                    arr = arr.astype(float)
+
+                if arr.ndim == 1 or (arr.ndim == 2 and arr.shape[1] in (3, 6, 9)):
+                    return arr
+
+            return data
+
+        if isinstance(data, int):
+            return float(data)
+
+        return data
+
+    if isinstance(data, np.ndarray):
         ret = data.tolist()
-        assert isinstance(ret, list)
+        assert isinstance(ret, (int, float, list))
         return ret
 
     if isinstance(data, Mapping):
@@ -55,7 +66,6 @@ def normalize(data: Entry, *, kind: Kind = Kind.DEFAULT) -> Entry:
         and len(data) <= 7
         and all(isinstance(d, (int, float)) for d in data)
     ):
-        data = cast(Sequence[float], data)
         return DimensionSet(*data)
 
     if isinstance(data, tuple) and kind == Kind.SINGLE_ENTRY and len(data) == 2:
@@ -65,17 +75,12 @@ def normalize(data: Entry, *, kind: Kind = Kind.DEFAULT) -> Entry:
     if is_sequence(data) and (kind == Kind.SINGLE_ENTRY or not isinstance(data, tuple)):
         return [normalize(d, kind=Kind.SINGLE_ENTRY) for d in data]
 
-    if isinstance(data, Dimensioned):
-        value = normalize(data.value, kind=Kind.SINGLE_ENTRY)
-        assert isinstance(value, (int, float, list))
-        return Dimensioned(value, data.dimensions, data.name)
-
     if isinstance(data, str):
         return parse_data(data)
 
     if isinstance(
         data,
-        (int, float, bool, tuple, DimensionSet),
+        (int, float, bool, tuple, DimensionSet, Dimensioned),
     ):
         return data
 
@@ -107,58 +112,35 @@ def dumps(
     if isinstance(data, DimensionSet):
         return b"[" + b" ".join(dumps(v) for v in data) + b"]"
 
-    if kind in (
-        Kind.ASCII_FIELD,
[…]
