+++ pyproject.toml
@@ -30,6 +30,8 @@ classifiers = [
 
 dependencies = [
     "aioshutil>=1,<2",
+    "numpy>=1.25.0,<3; python_version>='3.10'",
+    "numpy>=1,<3",
     "pyparsing>=3,<4",
     "typing-extensions>=4,<5; python_version<'3.11'",
 ]
@@ -37,13 +39,8 @@ dependencies = [
 dynamic = ["version"]
 
 [project.optional-dependencies]
-numpy = [
-    "numpy>=1.25.0,<3; python_version>='3.10'",
-    "numpy>=1,<3"
-]
 lint = ["ruff"]
 test = [
-    "foamlib[numpy]",
     "pytest>=7,<9",
     "pytest-asyncio>=0.21,<0.25",
     "pytest-cov",
@@ -53,12 +50,10 @@ typing = [
     "mypy>=1,<2",
 ]
 docs = [
-    "foamlib[numpy]",
     "sphinx>=5,<9",
     "sphinx_rtd_theme",
 ]
 dev = [
-    "foamlib[numpy]",
     "foamlib[lint]",
     "foamlib[test]",
     "foamlib[typing]",

[…]
