### 2024-12-06 144ab4b
Bump version to 0.8.1
