### 2024-12-13 #7499 @gerlero
Hi @mbarzegary!
For disclosure, while expecting a longer waitlist period, we've made some updates to the package and released new versions since our submission. If you prefer, we can quickly update the paper to align with the changes in the latest version (0.8.1), or you’re of course welcome to proceed with reviewing the paper as initially submitted, which as stated corresponds to version 0.7.0 of the package. Let us know what works best for you.
