### 2024-12-13 b7e8b6c
Update JOSS paper
