+++ paper.bib
@@ -17,6 +17,17 @@
   url = {https://github.com/python/mypy}
 }
 
+@article{numpy,
+  author = {Harris, Charles R and Millman, K Jarrod and Van Der Walt, St{\'e}fan J and Gommers, Ralf and Virtanen, Pauli and Cournapeau, David and Wieser, Eric and Taylor, Julian and Berg, Sebastian and Smith, Nathaniel J and others},
+  title = {Array programming with NumPy},
+  journal = {Nature},
+  volume = {585},
+  number = {7825},
+  pages = {357--362},
+  year = {2020},
+  doi = {10.1038/s41586-020-2649-2}
+}
+
 @misc{openfoam-app,
   author = {Gerlero, Gabriel S.},
   title = {{OpenFOAM.app}: Native {OpenFOAM} for {macOS}},
