+++ foamlib/_cases/_async.py
@@ -106,7 +106,7 @@ class AsyncFoamCase(FoamCaseRunBase):
     async def _rmtree(
         path: os.PathLike[str] | str, *, ignore_errors: bool = False
     ) -> None:
-        await aioshutil.rmtree(path, ignore_errors=ignore_errors)
+        await aioshutil.rmtree(path, ignore_errors=ignore_errors)  # type: ignore [call-arg]
 
     @staticmethod
     async def _copytree(
