### 2024-12-14 8e2b2ff
Bump version to 0.8.2
