### 2024-12-16 #7499 @mbarzegary
Hi @gerlero
Nice. I believe having the latest/updated version would be definitely better, so please update the paper to match the new changes, after which we will proceed with the submission.
