+++ .github/workflows/ci.yml
@@ -16,9 +16,9 @@ jobs:
     - name: Checkout
       uses: actions/checkout@v4
     - name: Lint with Ruff
-      uses: astral-sh/ruff-action@v2
+      uses: astral-sh/ruff-action@v3
     - name: Format with Ruff
-      uses: astral-sh/ruff-action@v2
+      uses: astral-sh/ruff-action@v3
       with:
         args: 'format --check'
 
