### 2024-12-27 4c61583
Fix parsing of complex schemes
