### 2024-12-27 e8f348e
Fix type errors
