### 2024-12-30 #7499 @gerlero
@mbarzegary I apologize for the delay. The paper has now been updated to match the latest version of the software (0.8.3).
