### 2025-01-06 #7499 @mbarzegary
Hi @akashdhruv, @AndreWeiner, @paugier, @Failxxx 👋
Considering that you have already published your software in JOSS, would any of you be willing to review this submission? As you know, we carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Paper: https://raw.githubusercontent.com/openjournals/joss-papers/joss.07499/joss.07499/10.21105.joss.07499.pdf
Software: https://github.com/gerlero/foamlib
