### 2025-01-08 #7499 @paugier
I react a bit late but I'm also available to review.

It seems that Foamlib is a very nice project.

Note that we wrote in my lab [Fluidsimfoam](https://foss.heptapod.net/fluiddyn/fluidsimfoam/) another Python package to work with OpenFOAM. In Fluidsimfoam, we also have a parser for OpenFOAM files (with [a lot of tests](https://foss.heptapod.net/fluiddyn/fluidsimfoam/-/tree/branch/default/tests)) and I can confirm that it is not simple to be right for all cases.

We also have in  https://github.com/fluiddyn/fluidfoam mostly to analyze OpenFOAM results.

These different Python tools can complement each other.
