### 2025-01-08 #7499 @mbarzegary
@akashdhruv, @AndreWeiner, @paugier, @Failxxx thank you guys for your help.

@paugier thanks for the information. Although having more than 3 reviewers is not very common at JOSS, I think having you on board will be beneficial for the current submission due to your experience with Python wrappers for OpenFOAM.
