### 2025-01-08 #7633 @editorialbot
[REVIEW]: foamlib: A modern Python package for working with OpenFOAM
**Submitting author:** @gerlero (<a href="http://orcid.org/0000-0002-5138-0328">Gabriel S. Gerlero</a>)
**Repository:** https://github.com/gerlero/foamlib
**Branch with paper.md** (empty if default branch): joss
**Version:** v1.0.0

@akashdhruv & @AndreWeiner & @Failxxx & @paugier, your review will be checklist based. Each of you will have a separate checklist that you should update when carrying out your review.
