### 2025-01-08 3d215dc
Handle #-directives as keywords in OpenFOAM files
