### 2025-01-27 gerlero/foamlib#317 @AndreWeiner
JOSS review - documentation
Hi @gerlero,
your documentation currently misses a few items to conform with the JOSS publication guidelines. Here are a couple of items to check:

1. The README should include a short statement of need, similar to what you wrote in the article.
2. Example usage: you included a couple of code snippets that demonstrate small parts of the API, which is great; however, it would be more useful to include at least one example of a more realistic use case (not necessarily in the README); the one-liners/minimal code snippets, you could also include in the API's docstrings ([see here](https://stackoverflow.com/questions/46892227/how-do-i-insert-highlight-or-code-block-into-sphinx-style-docstrings)).
3. the API has extensive type hinting, which is great; at least for the main classes (FoamCaseBase, FoamCase, Async*, FoamFile*) I suggest adding also some more elaborate docstrings describing the main purpose of the class, what it adds to a base class, what the current limitations are, and how to use it.
4. You set up automated testing; for developers/contributors, a small set of instructions on running unit tests locally would be helpful.
5. Currently, contributor guidelines are missing (refer to *community guidelines* in [this list](https://joss.readthedocs.io/en/latest/review_checklist.html#documentation))

@akashdhruv, @Failxxx, @paugier

Best, Andre
