### 2025-01-27 gerlero/foamlib#316 @gerlero
Thanks, @AndreWeiner! Heads up, I’m on vacation right now; I’ll be replying next week.
