### 2025-02-03 gerlero/foamlib#317 @Failxxx
Related to: openjournals/joss-reviews/issues/7633
