### 2025-02-03 gerlero/foamlib#316 @Failxxx
Related to: openjournals/joss-reviews/issues/7633
