### 2025-02-07 gerlero/foamlib#316 @gerlero
@AndreWeiner thanks again for the review! I'll reply to your comments one by one, listing the changes that I intend to make in the paper or docs (and then checking them off as I make them).

> 1. **line 20**: could you comment on why you decided to write a parser rather than create C++ bindings? Maybe a short discussion of the pros and cons of different approaches could be helpful for readers.

- [ ] PAPER: Comment on why an external parser was written instead of C++ bindings for OpenFOAM's own parser.

> 2. **line 66**: How is parallelism handled when creating a `FoamCaseBase` instance? Let's say I simulated in parallel and then reconstructed the time folders. Will `TimeDirectory` loop over the reconstructed or the decomposed time folders?

- [ ] PAPER: Mention that looping over a `TimeDirectory` will iterate over the reconstructed time folders.

> 3. **line 122**: How is the SLURM execution configured? I couldn't find any information in the documentation or the API.

- [ ] DOCS: describe how Slurm-based execution works with `AsyncSlurmFoamCase`.

> 4. **line 129**: I haven't tested the code yet, but I believe it is unlikely that the parser supports the full extent of OpenFOAM syntax. For example, have you considered regular expressions, macro expansions, or include directives ([overview](https://doc.cfd.direct/openfoam/user-guide-v10/basic-file-format))? These expressions are frequently encountered in all sorts of dictionaries. Maybe you have to soften _fully understands_ slightly. In general, I think it would be sensible to include a section with unsupported file format and other features in the README or the documentation and to refer to it in the article. For example, I also noticed that the current implementation for computing the cell centers will not work for distributed meshes, which is perfectly fine. Still, to avoid upsetting users, it's better to document it somewhere.

- [ ] DOCS: mention that the `FoamFile` class supports regular expressions, include directives, and other `#`-based expansions; but does not itself expand these expressions.
- [ ] DOCS: mention that `FoamFile` does not (currently) support `#codeStream`s.
- [ ] DOCS: mention that `TimeDirectory` cell center computation only works with reconstructed meshes.
- [ ] PAPER: soften language on statement of what files `FoamFile` can parse & refer readers to `FoamFile` API docs.

> 5. **line 131-132**: I guess `FoamFile` does not support outputs written to _postProcessing_ like force coefficients, probes, slices, etc. I suggest clarifying this point since users/readers might not know that not all OpenFOAM output files are `FoamFile`s.

- [ ] DOCS: mention that non-FoamFile output files are also not supported by the `FoamFile` class.

> 11. **table 1**: in the second column of _uniform field_, you write `np.ndarray` while in other lines, you use `numpy.ndarray`; if possible in terms of space, I suggest using either one or the other.

- [ ] PAPER: fix inconsistency `numpy.ndarray` vs `np.ndarray`.

I'll start work on making these modifications. Do let me know anytime if you find that any of the proposed changes don't adequately respond to your remarks.
