### 2025-02-07 gerlero/foamlib#317 @gerlero
@AndreWeiner @Failxxx Proposed changes:

> 1. The README should include a short statement of need, similar to what you wrote in the article.

- [ ] README: add statement of need.

> 2. Example usage: you included a couple of code snippets that demonstrate small parts of the API, which is great; however, it would be more useful to include at least one example of a more realistic use case (not necessarily in the README); the one-liners/minimal code snippets, you could also include in the API's docstrings ([see here](https://stackoverflow.com/questions/46892227/how-do-i-insert-highlight-or-code-block-into-sphinx-style-docstrings)).

- [ ] README or DOCS (not sure where yet): add a realistic, self-contained, use case example.
- [ ] DOCS: add code snippets from README to the relevant docstrings.

> 3. the API has extensive type hinting, which is great; at least for the main classes (FoamCaseBase, FoamCase, Async*, FoamFile*) I suggest adding also some more elaborate docstrings describing the main purpose of the class, what it adds to a base class, what the current limitations are, and how to use it.

- [ ] DOCS: add more elaborate docstrings for main user-facing classes.

> 4. You set up automated testing; for developers/contributors, a small set of instructions on running unit tests locally would be helpful.
> 5. Currently, contributor guidelines are missing (refer to _community guidelines_ in [this list](https://joss.readthedocs.io/en/latest/review_checklist.html#documentation))

- [ ] README: add clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
- [ ] README: in contributing section, mention how to run the tests locally.
