### 2025-02-13 #7633 @Failxxx
Hello @mbarzegary,
I think it is going well.
I am waiting for this issue [gerlero/foamlib#317](https://github.com/gerlero/foamlib/issues/317) to be completed.

I also need to take some time to read again the draft paper, see if I missed something that is not already mentioned on this issue [gerlero/foamlib#316](https://github.com/gerlero/foamlib/issues/316).

Sincerely,
Félix.
