### 2025-02-14 gerlero/foamlib#317 @gerlero
> Can you please link the PR to this issue once your changes are ready?

Of course. I’ll link the relevant PRs as I get to work on them.
