### 2025-03-03 gerlero/foamlib#317 @paugier
I agree that some work is needed for the documentation. The "Documentation" section of the review guidelines is indeed a good start. Some text of the article could also be used.

Also PyFoam is mentioned but I think other Python packages related to OpenFOAM should also be mentioned in the section Statement of Need, like fluidfoam and fluidsimfoam.
