### 2025-03-03 gerlero/foamlib#323 @paugier
Increase testing on parsing
It would be interesting to adapt the tests in

https://foss.heptapod.net/fluiddyn/fluidsimfoam/-/tree/branch/default/tests

for foamlib to know what is really supported by foamlib's parser and be able to be more precise in the article and the documentation.

Related to https://github.com/openjournals/joss-reviews/issues/7633
