### 2025-03-03 gerlero/foamlib#317 @gerlero
@paugier Proposed changes:

> I agree that some work is needed for the documentation. The "Documentation" section of the review guidelines is indeed a good start. Some text of the article could also be used.

- [ ] DOCS: improve documentation following the [Documentation section of the review guidelines](https://joss.readthedocs.io/en/latest/review_checklist.html#documentation) (consider extracting text from the paper).

> Also PyFoam is mentioned but I think other Python packages related to OpenFOAM should also be mentioned in the section Statement of Need, like fluidfoam and fluidsimfoam.

- [ ] PAPER: mention other OpenFOAM-related Python packages in the statement of need, including `fluidfoam` and `fluidsimfoam`.
