### 2025-03-03 gerlero/foamlib#323 @gerlero
@paugier Proposed changes:

- [ ] TESTS: add extra parsing tests based on `fluidsimfoam`'s tests at https://foss.heptapod.net/fluiddyn/fluidsimfoam/-/tree/branch/default/tests
