+++ foamlib/_files/_files.py
@@ -40,9 +40,46 @@ class FoamFile(
     """
     An OpenFOAM data file.
 
-    Use as a mutable mapping (i.e., like a dict) to access and modify entries.
+    `FoamFile` supports most OpenFOAM data and configuration files (i.e., files with a
+    "FoamFile" header), including those with regular expressions and #-based directives.
+    Notable exceptions are FoamFiles with #codeStreams and those multiple #-directives
+    with the same name, which are currently not supported. Non-FoamFile output files are
+    also not suppored by this class. Regular expressions and #-based directives can be
+    accessed and modified, but they are not evaluated or expanded by this library.
 
-    Use as a context manager to make multiple changes to the file while saving all changes only once at the end.
+    Use `FoamFile` as a mutable mapping (i.e., like a `dict`) to access and modify
+    entries. When accesing a sub-dictionary, the returned value will be a
+    `FoamFile.SubDict` object, that allows for further access and modification of nested
+    dictionaries within the `FoamFile` in a single operation.
+
+    If the `FoamFile` does not store a dictionary, the main stored value can be accessed
+    and modified by passing `None` as the key (e.g., `file[None]`).
+
+    You can also use the `FoamFile` as a context manager (i.e., within a `with` block)
+    to make multiple changes to the file while saving any and all changes only once at
+    the end.
+
+    :param path: The path to the file. If the file does not exist, it will be created
+        when the first change is made. However, if an attempt is made to access entries
+        in a non-existent file, a `FileNotFoundError` will be raised.
+
+    Example usage: ::
+        from foamlib import FoamFile
+
+        file = FoamFile("path/to/case/system/controlDict") # Load a controlDict file
+        print(file["endTime"]) # Print the end time
+        file["writeInterval"] = 100 # Set the write interval to 100
+        file["writeFormat"] = "binary" # Set the write format to binary
+
+    or (better): ::
+        from foamlib import FoamCase
+
+        case = FoamCase("path/to/case")
+
+        with case.control_dict as file: # Load the controlDict file
+            print(file["endTime"]) # Print the end time
+            file["writeInterval"] = 100
+            file["writeFormat"] = "binary"
     """
 
     Dimensioned = Dimensioned
@@ -51,7 +88,32 @@ class FoamFile(
     class SubDict(
         MutableMapping[str, MutableEntry],
     ):
-        """An OpenFOAM dictionary within a file as a mutable mapping."""
+        """
+        An OpenFOAM sub-dictionary within a file.
+
+        `FoamFile.SubDict` is a mutable mapping that allows for accessing and modifying
+        nested dictionaries within a `FoamFile` in a single operation. It behaves like a
+        `dict` and can be used to access and modify entries in the sub-dictionary.
+
+        To obtain a `FoamFile.SubDict` object, access a sub-dictionary in a `FoamFile`
+        object (e.g., `file["subDict"]`).
+
+        Example usage: ::
+            from foamlib import FoamFile
+
+            file = FoamFile("path/to/case/system/fvSchemes") # Load an fvSchemes file
+            print(file["ddtSchemes"]["default"]) # Print the default ddt scheme
+            file["ddtSchemes"]["default"] = "Euler" # Set the default ddt scheme
+
+        or (better): ::
+            from foamlib import FoamCase
+
+            case = FoamCase("path/to/case")
+
+            with case.fv_schemes as file: # Load the fvSchemes file
+                print(file["ddtSchemes"]["default"])
+                file["ddtSchemes"]["default"] = "Euler"
+        """
 
         def __init__(self, _file: FoamFile, _keywords: tuple[str, ...]) -> None:
             self._file = _file
[…]
