+++ pyproject.toml
@@ -69,10 +69,7 @@ Documentation = "https://foamlib.readthedocs.io"
 path = "foamlib/__init__.py"
 
 [tool.mypy]
-packages = [
-    "foamlib",
-    "tests",
-]
+packages = ["foamlib"]
 strict = true
 
 [tool.ruff.lint]
