+++ pyproject.toml
@@ -79,10 +79,7 @@ Documentation = "https://foamlib.readthedocs.io"
 path = "foamlib/__init__.py"
 
 [tool.mypy]
-packages = [
-    "foamlib",
-    "tests",
-]
+packages = ["foamlib"]
 strict = true
 
 [tool.ruff.lint]
