### 2025-03-05 #7633 @gerlero
Thanks again to the editor and all four reviewers. The first batch of revisions is ready for further review in the form of pull requests https://github.com/gerlero/foamlib/pull/325 and https://github.com/gerlero/foamlib/pull/326 in the project's repo. Not sure how you want to proceed, but in any case you are welcome to comment on those PRs or on the related issues.

I kindly ask for more time to make the other requested changes; my intent currently is to be making those revisions via https://github.com/gerlero/foamlib/pull/326 and https://github.com/gerlero/foamlib/pull/327.
