### 2025-03-09 03e043d
Bump version to 0.8.5
