+++ foamlib/__init__.py
@@ -1,6 +1,6 @@
 """A Python interface for interacting with OpenFOAM."""
 
-__version__ = "0.8.4"
+__version__ = "0.8.5"
 
 from ._cases import (
     AsyncFoamCase,
