### 2025-03-17 89bd539
Fix parsing of tokens starting with 'inf' or 'nan'
