+++ .github/workflows/ci.yml
@@ -49,11 +49,11 @@ jobs:
         options: --health-cmd="mysqladmin ping" --health-interval=10s --health-timeout=5s --health-retries=3
     strategy:
       matrix:
-        openfoam-version: [2406, 2006, 12, 9]
+        openfoam-version: [2412, 2006, 12, 9]
         python-version: ['3.7', '3.8', '3.9', '3.10', '3.11', '3.12', '3.13']
         slurm: [false]
         include:
-          - openfoam-version: 2406
+          - openfoam-version: 2412
             python-version: '3.13'
             slurm: true
           - openfoam-version: 12
