### 2025-03-17 f869866
Build Docker images with OpenFOAM v2412
