### 2025-03-17 1ca2197
Bump version to 0.8.6
