### 2025-03-24 e4fad63
Add progress indicator for runs
