### 2025-03-24 0338364
Fix shell invocation
