### 2025-03-24 697c11a
Fix mypy errors
