### 2025-03-25 bda5abf
Bump version to 0.8.7
