### 2025-03-27 e342d3a
Add test based on the example code
