+++ pyproject.toml
@@ -45,6 +45,7 @@ test = [
     "pytest>=7,<9",
     "pytest-asyncio>=0.21,<0.26",
     "pytest-cov",
+    "scipy>=1,<2",
 ]
 typing = [
     "foamlib[test]",
