### 2025-04-08 d8596ac
Update foamlib/_cases/_base.py
Co-authored-by: Félix <felix.olart56@gmail.com>
