### 2025-04-08 5c736a4
Update foamlib/_files/_files.py
Co-authored-by: Félix <felix.olart56@gmail.com>
