### 2025-04-08 54f7a81
Update foamlib/_cases/_base.py
Co-authored-by: Félix <felix.olart56@gmail.com>
