### 2025-04-16 fe436b2
Add basic CLI
