+++ pyproject.toml
@@ -33,6 +33,7 @@ dependencies = [
     "numpy>=1.25.0,<3; python_version>='3.10'",
     "numpy>=1,<3",
     "pyparsing>=3.1.2,<4",
+    "typer-slim>=0.13,<0.16",
     "typing-extensions>=4,<5; python_version<'3.11'",
     "rich>=13,<15",
 ]
@@ -65,6 +66,9 @@ dev = [
 Homepage = "https://github.com/gerlero/foamlib"
 Repository = "https://github.com/gerlero/foamlib"
 Documentation = "https://foamlib.readthedocs.io"
+
+[project.scripts]
+foamlib = "foamlib.__main__:app"
 
 [tool.hatch.version]
 path = "foamlib/__init__.py"
