### 2025-04-18 52b2319
Fix parsing of identifiers with parentheses
