+++ pyproject.toml
@@ -30,35 +30,23 @@ classifiers = [
 
 dependencies = [
     "aioshutil>=1,<2",
-    "numpy>=1.25.0,<3; python_version>='3.10'",
     "numpy>=1,<3",
     "pyparsing>=3.1.2,<4",
-    "typing-extensions>=4,<5; python_version<'3.11'",
     "rich>=13,<15",
+    "typing-extensions>=4,<5; python_version<'3.11'",
 ]
 
 dynamic = ["version"]
 
-[project.optional-dependencies]
-lint = ["ruff"]
-test = [
+[dependency-groups]
+dev = [
+    "mypy>=1,<2",
     "pytest>=7,<9",
     "pytest-asyncio>=0.21,<0.27",
     "pytest-cov",
-]
-typing = [
-    "foamlib[test]",
-    "mypy>=1,<2",
-]
-docs = [
+    "ruff",
     "sphinx>=5,<9",
     "sphinx_rtd_theme",
-]
-dev = [
-    "foamlib[lint]",
-    "foamlib[test]",
-    "foamlib[typing]",
-    "foamlib[docs]",
 ]
 
 [project.urls]
