### 2025-04-19 6f3618c
Improve type hints
