### 2025-04-23 #7633 @paugier
I see that the PR related to this review have not yet been merged. Could they be merged? It would be nice to be able to read the new README and doc from the official website.

I'm happy with the inclusion of the tests from fluidsimfoam.

Sorry for this silly question but where can I read the new version of the manuscript? I can see the PR (https://github.com/gerlero/foamlib/pull/324) but is it possible to see the result?
