+++ tests/test_example.py
@@ -3,7 +3,7 @@ from pathlib import Path
 import numpy as np
 import pytest
 from foamlib import FoamCase
-from scipy.special import erfc  # type: ignore [import-untyped]
+from scipy.special import erfc
 
 
 def test_example(tmp_path: Path) -> None:
@@ -108,7 +108,7 @@ def test_example(tmp_path: Path) -> None:
 
     with case[0]["U"] as f:
         f.dimensions = f.DimensionSet(length=1, time=-1)
-        f.internal_field = [1, 0, 0]
+        f.internal_field = [1, 0, 0]  # type: ignore [assignment]
         f.boundary_field = {  # type: ignore [assignment]
             "inletUp": {"type": "fixedValue", "value": [1, 0, 0]},
             "inletDown": {"type": "fixedValue", "value": [1, 0, 0]},
