### 2025-04-23 #7633 @gerlero
> I see that the PR related to this review have not yet been merged. Could they be merged? It would be nice to be able to read the new README and doc from the official website.

@paugier Done, merged!

> I'm happy with the inclusion of the tests from fluidsimfoam.

Thanks again for writing and suggesting them.

> Sorry for this silly question but where can I read the new version of the manuscript? I can see the PR ([gerlero/foamlib#324](https://github.com/gerlero/foamlib/pull/324)) but is it possible to see the result?

https://github.com/gerlero/foamlib/actions/runs/14622601739/artifacts/2995314783
