### 2025-04-24 #7633 @paugier
I think the new version is suitable for publication in JOSS.

Just a last remark: a bit unfortunate that https://foamlib.readthedocs.io/en/latest/index.html is much less nice that the README!
