### 2025-04-28 23f9bae
Bump foamlib from 0.8.1 to 0.8.8
Bumps [foamlib](https://github.com/gerlero/foamlib) from 0.8.1 to 0.8.8.
- [Release notes](https://github.com/gerlero/foamlib/releases)
- [Commits](https://github.com/gerlero/foamlib/compare/v0.8.1...v0.8.8)

---
updated-dependencies:
- dependency-name: foamlib
  dependency-version: 0.8.8
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>
