### 2025-04-30 bf630f7
ENH: add table reader with various extensions
