+++ pyproject.toml
@@ -31,6 +31,7 @@ classifiers = [
 dependencies = [
     "aioshutil>=1,<2",
     "numpy>=1,<3",
+    "pandas>=1,<3",
     "pyparsing>=3.1.2,<4",
     "rich>=13,<15",
     "typing-extensions>=4,<5; python_version<'3.11'",
