### 2025-04-30 b9fdd96
ENH: extract column names from file
