+++ tests/test_postprocessing/test_load_output_files.py
@@ -1,4 +1,36 @@
-from foamlib.postprocessing.tableReader import TableReader
+from foamlib.postprocessing.tableReader import TableReader, extract_column_names
+
+
+def test_read_headers() -> None:
+    force_file = "tests/test_postprocessing/postProcessing/forces/0/force.dat"
+    force_headers = extract_column_names(force_file)
+    assert force_headers == ["Time", "total_x", "total_y", "total_z", "pressure_x", "pressure_y", "pressure_z", "viscous_x", "viscous_y", "viscous_z"]
+
+    surface_p_file = "tests/test_postprocessing/postProcessing/freeSurface/0.1/p_freeSurface.raw"
+    surface_p_headers = extract_column_names(surface_p_file)
+    assert surface_p_headers == ["x", "y", "z", "p"]
+
+    surface_u_file = "tests/test_postprocessing/postProcessing/freeSurface/0.1/U_freeSurface.raw"
+    surface_u_headers = extract_column_names(surface_u_file)
+    assert surface_u_headers == ["x", "y", "z", "U_x", "U_y", "U_z"]
+
+    probes_p_file = "tests/test_postprocessing/postProcessing/probes/0/p"
+    probes_p_headers = extract_column_names(probes_p_file)
+    assert probes_p_headers == ["Time", "0"]
+
+    probes_u_file = "tests/test_postprocessing/postProcessing/probes/0/U"
+    probes_u_headers = extract_column_names(probes_u_file)
+    assert probes_u_headers == ["Time", "0"]
+
+    probes_T_file = "tests/test_postprocessing/postProcessing/probes/0/T"
+    probes_T_headers = extract_column_names(probes_T_file)
+    assert probes_T_headers == ["Time", "0"]
+
+    sample_xy_file = "tests/test_postprocessing/postProcessing/sample1/0.1/centreLine_T.xy"
+    sample_xy_headers = extract_column_names(sample_xy_file)
+    assert sample_xy_headers == None
+
+    
 
 def test_read_forces() -> None:
     reader = TableReader()
