### 2025-05-01 4cf6f55
simplify tests
