### 2025-05-01 516691f
applied all fixes from ruff
