+++ foamlib/postprocessing/table_reader.py
@@ -1,6 +1,10 @@
+"""This module provides a utility class for reading tabular data from files with various extensions."""
+
+from __future__ import annotations  # Add this import at the top of the file
+
 from itertools import islice
 from pathlib import Path
-from typing import Callable, Dict, Optional, List
+from typing import Callable, ClassVar, Optional
 
 import pandas as pd
 
@@ -8,6 +12,7 @@ import pandas as pd
 class TableReader:
     """
     TableReader is a utility class for reading tabular data from files with various extensions.
+
     It uses a registry pattern to associate file extensions with specific reader functions.
 
     Attributes:
@@ -26,65 +31,85 @@ class TableReader:
             for the file's extension.
     """
 
-    _registry: Dict[str, Callable[[str], pd.DataFrame]] = {}
+    _registry: ClassVar[dict[str, Callable[[str], pd.DataFrame]]] = {}
 
-    def __init__(self):
-        pass
+
+    def __init__(self)-> None:
+        """Initialize the TableReader instance."""
+
 
     @classmethod
-    def register(cls, extension: str):
+    def register(cls, extension: str) -> Callable[[Callable[[str], pd.DataFrame]], Callable[[str], pd.DataFrame]]:
         """
-        A class method decorator used to register a reader function for a specific file extension.
+        Register a reader function for a specific file extension.
+
         The extension is case-insensitive.
+
         Args:
             extension (str): The file extension (e.g., ".dat", ".raw") to register the reader for.
+
         Returns:
             Callable[[Callable[[str], pd.DataFrame]], Callable[[str], pd.DataFrame]]:
                 A decorator that registers the function as a reader for the specified extension.
         """
-
-        def decorator(func: Callable[[str], pd.DataFrame]):
+        def decorator(func: Callable[[str], pd.DataFrame]) -> Callable[[str], pd.DataFrame]:
             cls._registry[extension.lower()] = func
             return func
 
         return decorator
 
     def read(
-        self, filepath: str, column_names: Optional[List[str]] = None
+        self, filepath: str, column_names: Optional[list[str]] = None
     ) -> pd.DataFrame:
         """
-        Reads a file and returns its contents as a pandas DataFrame.
+        Read a file and return its contents as a pandas DataFrame.
+
         The file extension is used to determine the appropriate reader function.
+
         Raises:
             ValueError: If no reader is registered for the file's extension.
+
         Args:
             filepath (str): The path to the file to be read.
+
         Returns:
             pd.DataFrame: The contents of the file as a pandas DataFrame.
         """
         ext = Path(filepath).suffix.lower()
         if ext not in self._registry:
-            raise ValueError(f"No reader registered for extension: '{ext}'")
+            error_message = f"No reader registered for extension: '{ext}'"
+            raise ValueError(error_message)
         return self._registry[ext](filepath, column_names=column_names)
 
 
-def is_convertible_to_float(values: List[str]) -> bool:
+def is_convertible_to_float(values: list[str]) -> bool:
     """
     Check if all values in a list are convertible to floats.
+
     Args:
-        values (List[str]): A list of string values to check.
+        values (list[str]): A list of string values to check.
     Returns:
         bool: True if all values are convertible to floats, False otherwise.
     """
     try:
         [float(value) for value in values]
-        return True
     except ValueError:
         return False
+    else:
+        return True
 
 
-def extract_column_names(filepath: str) -> Optional[List[str]]:
-    with open(filepath, "r") as f:
[…]
