+++ pyproject.toml
@@ -103,6 +103,7 @@ extend-ignore = [
     "S603",
     "SLF001",
     "TID252",
+    "UP007"
 ]
 
 [tool.ruff.lint.pydocstyle]
