### 2025-05-01 6683a31
fixes ruff suggestions in test
