### 2025-05-01 c3ae493
FIX: Path.open is not supported by older versions
