### 2025-05-01 e5969ab
ENH: get openfoam case in folder
