### 2025-05-05 4332a9c
Add special case for bool-like keywords
