+++ foamlib/__init__.py
@@ -1,6 +1,13 @@
 """A Python interface for interacting with OpenFOAM."""
 
-__version__ = "0.8.10"
+import sys
+
+if sys.version_info >= (3, 8):
+    import importlib.metadata as importlib_metadata
+else:
+    import importlib_metadata
+
+__version__ = importlib_metadata.version("foamlib")
 
 from ._cases import (
     AsyncFoamCase,
