+++ pyproject.toml
@@ -1,9 +1,10 @@
 [build-system]
-requires = ["hatchling"]
-build-backend = "hatchling.build"
+requires = ["uv_build>=0.7.2,<0.8"]
+build-backend = "uv_build"
 
 [project]
 name = "foamlib"
+version = "0.8.10"
 description = "A Python interface for interacting with OpenFOAM"
 readme = "README.md"
 requires-python = ">=3.7"
@@ -30,13 +31,15 @@ classifiers = [
 
 dependencies = [
     "aioshutil>=1,<2",
+    "importlib_metadata>=1.4,<8; python_version<'3.8'",
     "numpy>=1,<3",
     "pyparsing>=3.1.2,<4",
     "rich>=13,<15",
     "typing-extensions>=4,<5; python_version<'3.11'",
 ]
 
-dynamic = ["version"]
+[tool.uv.build-backend]
+module-root = ""
 
 [dependency-groups]
 dev = ["foamlib[dev]"]
@@ -75,9 +78,6 @@ dev = [
 Homepage = "https://github.com/gerlero/foamlib"
 Repository = "https://github.com/gerlero/foamlib"
 Documentation = "https://foamlib.readthedocs.io"
-
-[tool.hatch.version]
-path = "foamlib/__init__.py"
 
 [tool.mypy]
 packages = ["foamlib"]
