### 2025-05-06 d082925
Remove unneeded extras
