### 2025-05-06 e359fe0
Re-enable type checking of tests
