### 2025-05-06 473437f
Bump version to 0.8.11
