### 2025-05-08 b1f16ee
Fast parse certain numeric lists to NumPy arrays
