+++ tests/test_files/test_parsing/test_poly_mesh.py
@@ -2,6 +2,7 @@
 
 from pathlib import Path
 
+import numpy as np
 from foamlib import FoamFile
 
 contents = r"""
@@ -49,18 +50,18 @@ def test_get_cells_coords(tmp_path: Path) -> None:
     file = FoamFile(path)
 
     points = file[None]
-    assert isinstance(points, list)
+    assert isinstance(points, np.ndarray)
 
-    assert points[0] == [0, 0, 0]
-    assert points[1] == [0.15707963268, 0, 0]
-    assert points[2] == [0.314159265359, 0, 0]
-    assert points[3] == [0.471238898038, 0, 0]
-    assert points[4] == [0.628318530718, 0, 0]
-    assert points[5] == [0, 0, 0]
-    assert points[6] == [0.15707963268, 0, 0]
-    assert points[7] == [0.314159265359, 0, 0]
-    assert points[8] == [0.471238898038, 0, 0]
-    assert points[9] == [0.628318530718, 0, 0]
+    assert np.array_equal(points[0], [0, 0, 0])
+    assert np.array_equal(points[1], [0.15707963268, 0, 0])
+    assert np.array_equal(points[2], [0.314159265359, 0, 0])
+    assert np.array_equal(points[3], [0.471238898038, 0, 0])
+    assert np.array_equal(points[4], [0.628318530718, 0, 0])
+    assert np.array_equal(points[5], [0, 0, 0])
+    assert np.array_equal(points[6], [0.15707963268, 0, 0])
+    assert np.array_equal(points[7], [0.314159265359, 0, 0])
+    assert np.array_equal(points[8], [0.471238898038, 0, 0])
+    assert np.array_equal(points[9], [0.628318530718, 0, 0])
 
     assert len(points) == 10
 
