### 2025-05-08 de7b9b5
Remove unused internal enum
