### 2025-05-08 921682a
WIP:. added functionality to load combines tables from a folder
