+++ foamlib/postprocessing/load_tables.py
@@ -1,6 +1,12 @@
 from __future__ import annotations  # Add this import at the top of the file
 
+from dataclasses import dataclass, field
 import os
+import pandas as pd
+from pathlib import Path
+from typing import Union
+from collections import defaultdict
+from .table_reader import TableReader
 
 
 def _of_case(dirnames: list[str], filenames: list[str]) -> bool:
@@ -17,8 +23,9 @@ def _of_case(dirnames: list[str], filenames: list[str]) -> bool:
     ofcase : bool
         is the folder an OpenFOAM Case
     """
-    return "constant" in dirnames and "system" in dirnames
-
+    hasConstant = "constant" in dirnames
+    hasSystem = "system" in dirnames
+    return hasConstant and hasSystem
 
 
 def of_cases(dir_name: str) -> list[str]:
@@ -39,3 +46,109 @@ def of_cases(dir_name: str) -> list[str]:
             cases.append(path)
             dirnames[:] = []
     return cases
+
+
+@dataclass
+class OutputFile:
+    file_name: str
+    folder: Union[str, Path] = None
+    _times: set[float] = field(default_factory=set, init=False, repr=False)
+
+    def add_time(self, t: float):
+        self._times.add(t)
+
+    @property
+    def times(self) -> list[float]:
+        return sorted(self._times)
+
+
+def load_tables(output_file: OutputFile, dir_name: str) -> Optional[pd.DataFrame]:
+    """
+    Load and concatenate all available dataframes for an OutputFile across time steps.
+
+    Parameters
+    ----------
+    output_file : OutputFile
+        OutputFile object containing file name, time steps, and folder
+    dir_name : str
+        Root directory where OpenFOAM cases are stored
+
+    Returns
+    -------
+    pd.DataFrame or None
+        Concatenated dataframe of all available time steps, or None if nothing found
+    """
+    all_tables = []
+
+    for case in of_cases(dir_name):
+        case_path = Path(case)
+        postproc_root = case_path / "postProcessing"
+
+        if not postproc_root.exists():
+            continue
+
+        for time in output_file.times:
+            file_path = (
+                postproc_root / output_file.folder / str(time) / output_file.file_name
+            )
+
+            if file_path.exists():
+                print(f"Loading {file_path}")
+                reader = TableReader()
+                table = reader.read(file_path)
+                file_path
+                table["__case__"] = case_path.name
+                table["__time__"] = time
+                all_tables.append(table)
+
+    if all_tables:
+        return pd.concat(all_tables, ignore_index=True)
+    else:
+        print(f"No data found for {output_file.file_name}")
+        return None
+
+
+def is_float(s: str) -> bool:
+    try:
+        float(s)
+    except ValueError:
+        return False
+    return True
+
+
+def _outputfiles(file_map, case_path_str, postproc_root):
+    for dirpath, _, filenames in os.walk(postproc_root):
+        base = os.path.basename(dirpath)
+
+        if not is_float(base):
+            # Skip directories that are not time directories
+            continue
+
+        time = float(base)
+        time_path = Path(dirpath)
+        rel_to_postproc = time_path.relative_to(postproc_root)
+        folder = rel_to_postproc.parent
+        folder_str = str(folder) if folder != Path(".") else ""
+
+        for fname in filenames:
+            key = f"{folder_str}--{fname}"
+            full_folder = os.path.join(case_path_str, "postProcessing", folder)
+
+            if key not in file_map:
+                file_map[key] = OutputFile(file_name=fname, folder=full_folder)
+
+            file_map[key].add_time(time)
+
+
+def list_outputfiles(Cases: str = "Cases") -> dict[str, list[OutputFile]]:
[…]
