+++ tests/test_files/test_parsing/test_decompose_par.py
@@ -44,8 +44,7 @@ def test_simple(tmp_path: Path) -> None:
 
     assert decompose_par_dict["numberOfSubdomains"] == 8
     assert decompose_par_dict["method"] == "simple"
-    assert isinstance(decompose_par_dict["coeffs", "n"], np.ndarray)
-    assert np.array_equal(decompose_par_dict["coeffs", "n"], [4, 2, 1])
+    assert np.array_equal(decompose_par_dict["coeffs", "n"], [4, 2, 1])  # type: ignore[arg-type]
     assert decompose_par_dict["coeffs", "order"] == "xyz"
     assert decompose_par_dict["coeffs", "delta"] == 0.001
 
