### 2025-05-10 17627b3
fix ruff and mypy comments
