+++ tests/test_postprocessing/test_load_tables.py
@@ -1,16 +1,15 @@
 from __future__ import annotations
 
-import pytest
 import pandas as pd
 from foamlib.postprocessing.load_tables import (
-    of_cases,
-    load_tables,
-    list_outputfiles,
     OutputFile,
+    list_outputfiles,
+    load_tables,
+    of_cases,
 )
 
 
-def max_height_filter(table: pd.DataFrame,parameters: dict[str,str]) -> pd.DataFrame:
+def max_height_filter(table: pd.DataFrame, parameters: dict[str, str]) -> pd.DataFrame:
     d = {
         "x": [table["x"].max()],
         "y": [table["y"].max()],
@@ -31,11 +30,8 @@ def test_load_tables_forces() -> None:
     """Test if the load_tables function works correctly."""
 
     file = OutputFile(file_name="force.dat", folder="forces")
-    # file.add_time(0)
 
     table = load_tables(output_file=file, dir_name="tests/test_postprocessing/Cases")
-
-    
 
     assert table.columns.tolist()[-2:] == ["grid", "initHeight"]
     assert list(table["grid"].unique()) == ["res1", "res2", "res3"]
@@ -49,10 +45,12 @@ def test_load_tables_surface() -> None:
     file = OutputFile(file_name="U_freeSurface.raw", folder="freeSurface")
 
     table = load_tables(
-        output_file=file, dir_name="tests/test_postprocessing/Cases", filter=max_height_filter
+        output_file=file,
+        dir_name="tests/test_postprocessing/Cases",
+        filter_table=max_height_filter,
     )
 
-    assert table.columns.tolist() == ["x", "y", "z","grid", "initHeight", "timeValue"]
+    assert table.columns.tolist() == ["x", "y", "z", "grid", "initHeight", "timeValue"]
     assert list(table["grid"].unique()) == ["res1", "res2", "res3"]
     assert list(table["timeValue"].unique()) == [0.1, 0.2, 0.4]
     assert list(table["initHeight"].unique()) == ["height_02", "height_03"]
@@ -64,8 +62,8 @@ def test_output_files() -> None:
 
     output_files = list_outputfiles("tests/test_postprocessing/Cases")
     # output_files is a dictionary with keys as file names and values as OutputFile objects
-    # key: freeSurface--U_freeSurface.raw
-    folders = [key.split("--")[0] for key in output_files.keys()]
+
+    folders = [key.split("--")[0] for key in output_files]
     assert sorted(folders) == sorted(
         [
             "freeSurface",
@@ -80,7 +78,7 @@ def test_output_files() -> None:
         ]
     )
 
-    files = [key.split("--")[1] for key in output_files.keys()]
+    files = [key.split("--")[1] for key in output_files]
     assert sorted(files) == sorted(
         [
             "U_freeSurface.raw",
