### 2025-05-11 ec8c727
Factor out serialization logic
