+++ foamlib/_files/_serialization.py
@@ -1,7 +1,6 @@
 from __future__ import annotations
 
 import sys
-from enum import Enum, auto
 from typing import overload
 
 if sys.version_info >= (3, 9):
@@ -23,31 +22,32 @@ from ._types import (
 )
 
 
-class Kind(Enum):
-    DEFAULT = auto()
-    SINGLE_ENTRY = auto()
-    ASCII_FIELD = auto()
-    SCALAR_ASCII_FIELD = auto()
-    BINARY_FIELD = auto()
-    SCALAR_BINARY_FIELD = auto()
-    DIMENSIONS = auto()
-    KEYWORD = auto()
+@overload
+def normalize_data(
+    data: DataLike, *, keywords: tuple[str, ...] | None = None
+) -> Data: ...
 
 
 @overload
-def normalize(data: DataLike, *, kind: Kind = Kind.DEFAULT) -> Data: ...
+def normalize_data(
+    data: EntryLike, *, keywords: tuple[str, ...] | None = None
+) -> Entry: ...
 
 
-@overload
-def normalize(data: EntryLike, *, kind: Kind = Kind.DEFAULT) -> Entry: ...
-
-
-def normalize(data: EntryLike, *, kind: Kind = Kind.DEFAULT) -> Entry:
-    if kind in (
-        Kind.ASCII_FIELD,
-        Kind.SCALAR_ASCII_FIELD,
-        Kind.BINARY_FIELD,
-        Kind.SCALAR_BINARY_FIELD,
+def normalize_data(
+    data: EntryLike, *, keywords: tuple[str, ...] | None = None
+) -> Entry:
+    if keywords is not None and (
+        keywords == ("internalField",)
+        or (
+            len(keywords) == 3
+            and keywords[0] == "boundaryField"
+            and (
+                keywords[2] == "value"
+                or keywords[2] == "gradient"
+                or keywords[2].endswith(("Value", "Gradient"))
+            )
+        )
     ):
         if is_sequence(data):
             try:
@@ -61,41 +61,48 @@ def normalize(data: EntryLike, *, kind: Kind = Kind.DEFAULT) -> Entry:
                 if arr.ndim == 1 or (arr.ndim == 2 and arr.shape[1] in (3, 6, 9)):
                     return arr  # type: ignore [return-value]
 
-            return [normalize(d, kind=Kind.SINGLE_ENTRY) for d in data]
+            return [normalize_data(d) for d in data]
 
         if isinstance(data, int):
             return float(data)
 
-        return normalize(data)
+        return normalize_data(data)
+
+    if isinstance(data, Mapping):
+        return {normalize_keyword(k): normalize_data(v) for k, v in data.items()}
 
     if isinstance(data, np.ndarray):
         ret = data.tolist()
         assert isinstance(ret, (int, float, list))
         return ret
 
-    if isinstance(data, Mapping):
-        return {k: normalize(v, kind=kind) for k, v in data.items()}
-
     if (
-        kind == Kind.DIMENSIONS
+        not isinstance(data, DimensionSet)
+        and keywords is not None
+        and keywords == ("dimensions",)
         and is_sequence(data)
         and len(data) <= 7
         and all(isinstance(d, (int, float)) for d in data)
     ):
         return DimensionSet(*data)
 
-    if isinstance(data, tuple) and kind == Kind.SINGLE_ENTRY and len(data) == 2:
+    if keywords is None and isinstance(data, tuple) and len(data) == 2:
         k, v = data
-        return (normalize(k), normalize(v))
+        assert not isinstance(k, Mapping)
+        return (
+            normalize_keyword(k),
+            normalize_data(v) if not isinstance(v, Mapping) else v,
+        )  # type: ignore [return-value]
 
[…]
