### 2025-05-12 b12e5aa
Add dumps staticmethod to FoamFile
