### 2025-05-12 ff5ed37
Add loads staticmethod to FoamFile
