### 2025-05-12 5e86410
Improve FoamFile.DimensionSet and FoamFile.Dimensions classes
