### 2025-05-12 c41da36
Fix FoamFile loads parsing of keyword entries without values
