+++ foamlib/__init__.py
@@ -1,6 +1,6 @@
 """A Python interface for interacting with OpenFOAM."""
 
-__version__ = "0.9.2"
+__version__ = "0.9.3"
 
 from ._cases import (
     AsyncFoamCase,
