### 2025-05-12 ee885df
Bump foamlib from 0.8.1 to 0.9.3
Bumps [foamlib](https://github.com/gerlero/foamlib) from 0.8.1 to 0.9.3.
- [Release notes](https://github.com/gerlero/foamlib/releases)
- [Commits](https://github.com/gerlero/foamlib/compare/v0.8.1...v0.9.3)

---
updated-dependencies:
- dependency-name: foamlib
  dependency-version: 0.9.3
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
