### 2025-05-12 b56e15e
Update FoamFile.Dimensioned class
