### 2025-05-13 38e1b6c
Improve type hints
