+++ tests/test_files/test_parsing/test_intermediate.py
@@ -157,10 +157,10 @@ def test_list_with_dict() -> None:
     assert len(boundary) == 1
 
     upper_boundary = boundary[0]
-    assert upper_boundary[0] == "upperBoundary"
-    assert upper_boundary[1]["type"] == "cyclic"
-    assert upper_boundary[1]["neighbourPatch"] == "lowerBoundary"
-    assert np.array_equal(upper_boundary[1]["faces"], [[3, 7, 6, 2]])
+    assert upper_boundary[0] == "upperBoundary"  # type: ignore[index]
+    assert upper_boundary[1]["type"] == "cyclic"  # type: ignore[index, call-overload]
+    assert upper_boundary[1]["neighbourPatch"] == "lowerBoundary"  # type: ignore[index, call-overload]
+    assert np.array_equal(upper_boundary[1]["faces"], [[3, 7, 6, 2]])  # type: ignore[arg-type, index, call-overload]
 
 
 def test_list_with_str() -> None:
@@ -174,10 +174,10 @@ def test_list_with_str() -> None:
     assert len(blocks) == 5
 
     assert blocks[0] == "hex"
-    assert np.array_equal(blocks[1], [0, 1, 2, 3, 4, 5, 6, 7])
-    assert np.array_equal(blocks[2], [40, 40, 40])
+    assert np.array_equal(blocks[1], [0, 1, 2, 3, 4, 5, 6, 7])  # type: ignore[arg-type]
+    assert np.array_equal(blocks[2], [40, 40, 40])  # type: ignore[arg-type]
     assert blocks[3] == "simpleGrading"
-    assert np.array_equal(blocks[4], [1, 1, 1])
+    assert np.array_equal(blocks[4], [1, 1, 1])  # type: ignore[arg-type]
 
 
 def test_file_simple() -> None:
@@ -457,7 +457,7 @@ def test_for_blockmesh() -> None:
     assert len(blocks) == 4
 
     assert blocks[0] == "hex"
-    assert np.array_equal(blocks[1], [4, 6, 14, 12, 0, 2, 10, 8])
+    assert np.array_equal(blocks[1], [4, 6, 14, 12, 0, 2, 10, 8])  # type: ignore[arg-type]
     assert blocks[2] == [1, "$upstreamCells", "$cylinderBoxCells"]
     assert blocks[3] == "$expandBlock"
 
@@ -487,27 +487,27 @@ def test_blocks() -> None:
     assert len(blocks) == 22
 
     assert blocks[0] == "hex"
-    assert np.array_equal(blocks[1], [0, 1, 2, 3, 4, 5, 6, 7])
+    assert np.array_equal(blocks[1], [0, 1, 2, 3, 4, 5, 6, 7])  # type: ignore[arg-type]
     assert blocks[2] == "inletChannel"
-    assert np.array_equal(blocks[3], [40, 1, 64])
+    assert np.array_equal(blocks[3], [40, 1, 64])  # type: ignore[arg-type]
     assert blocks[4] == "simpleGrading"
-    assert np.array_equal(blocks[5], [1, 1, 1])
+    assert np.array_equal(blocks[5], [1, 1, 1])  # type: ignore[arg-type]
     assert blocks[6] == "hex"
-    assert np.array_equal(blocks[7], [4, 5, 6, 7, 8, 9, 10, 11, 12])
+    assert np.array_equal(blocks[7], [4, 5, 6, 7, 8, 9, 10, 11, 12])  # type: ignore[arg-type]
     assert blocks[8] == "inletChannel"
-    assert np.array_equal(blocks[9], [40, 1, 16])
+    assert np.array_equal(blocks[9], [40, 1, 16])  # type: ignore[arg-type]
[…]
