+++ tests/test_files/test_dumps.py
@@ -92,7 +92,6 @@ def test_serialize_data() -> None:
     assert dumps([("a", "b"), ("c", "d")]) == b"(a b; c d;)"
     assert dumps([("a", {"b": "c"}), ("d", {"e": "g"})]) == b"(a {b c;} d {e g;})"
     assert dumps([("a", [0, 1, 2]), ("b", {})]) == b"(a (0 1 2); b {})"
-    assert dumps([{"a": "b", "c": "d"}, {"e": "g"}]) == b"({a b; c d;} {e g;})"
     assert dumps(["water", "oil", "mercury", "air"]) == b"(water oil mercury air)"
     assert dumps("div(phi,U)") == b"div(phi,U)"
 
@@ -104,8 +103,8 @@ def test_serialize_file() -> None:
         == b"{FoamFile {version 2.0; format ascii; class dictionary;}} 1.0"
     )
     assert (
-        FoamFile.dumps([{"a": "b", "c": "d"}, {"e": "g"}])
-        == b"{FoamFile {version 2.0; format ascii; class dictionary;}} ({a b; c d;} {e g;})"
+        FoamFile.dumps({"a": "b", "c": "d"})
+        == b"{FoamFile {version 2.0; format ascii; class dictionary;}} a b; c d;"
     )
     assert (
         FoamFile.dumps({"internalField": [[1, 2, 3], [4, 5, 6]]})
