### 2025-05-13 6eb7a35
Improve type hints
