### 2025-05-13 1ae6c22
Improve serialization
