### 2025-05-13 ade5ead
Improve type hints
