+++ tests/test_files/test_parsing/test_loads.py
@@ -71,3 +71,6 @@ def test_loads() -> None:
     assert FoamFile.loads("a {b c; d e;}") == {"a": {"b": "c", "d": "e"}}
     assert FoamFile.loads("(a b; c d;)") == [("a", "b"), ("c", "d")]
     assert FoamFile.loads("keyword;") == {"keyword": ""}
+    assert FoamFile.loads("#include $FOAM_CASE/simControls") == {
+        "#include": "$FOAM_CASE/simControls"
+    }
