+++ docs/index.rst
@@ -1,9 +1,18 @@
-Documentation for `foamlib <https://github.com/gerlero/foamlib>`_
-=================================================================
+.. image:: https://github.com/gerlero/foamlib/blob/main/logo.png?raw=true
+   :alt: foamlib logo
+   :width: 200 px
+   :target: https://github.com/gerlero/foamlib
+=========================================================================
+
+`foamlib <https://github.com/gerlero/foamlib>`_ is a modern Python package that simplifies working with OpenFOAM cases and files by providing a standalone parser for seamless interaction with OpenFOAM's input/output data. It includes robust case-handling capabilities that reduce boilerplate code and enable efficient pre-processing, post-processing, and simulation management directly from Python. With support for ASCII- and binary-formatted fields, a fully type-hinted API, and asynchronous operations, foamlib offers a streamlined, Pythonic approach to automating and managing OpenFOAM workflows.
+
+Documentation
+=============
 
 .. toctree::
    :maxdepth: 2
 
+   example
    cases
    files
 
