### 2025-05-14 #7633 @gerlero
@mbarzegary I've improved the docs. Specifically, I have:

- Improved the homepage of the docs, adding the logo at the top and a short description of the library
- Moved the self-contained example snippet from the README (where it was hidden by the default) to now being first section of the docs
- Fixed many instances of docstrings not being displayed properly on the site

The changes are now live @ https://foamlib.readthedocs.io/en/latest/

@paugier Let me know if the docs site is now more like you'd like it to be, and if there's another action you'd want me to take regarding this.
