### 2025-05-14 0603a62
Improve parsing of face lists
