+++ foamlib/_files/_types.py
@@ -1,14 +1,25 @@
 from __future__ import annotations
 
 import sys
-from typing import Any, Dict, List, NamedTuple, Optional, Tuple, Union
+from typing import (
+    Any,
+    Dict,
+    Generic,
+    List,
+    NamedTuple,
+    Optional,
+    Tuple,
+    TypeVar,
+    Union,
+    overload,
+)
 
 import numpy as np
 
 if sys.version_info >= (3, 9):
-    from collections.abc import Mapping, MutableMapping, Sequence
+    from collections.abc import Iterator, Mapping, MutableMapping, Sequence
 else:
-    from typing import Mapping, MutableMapping, Sequence
+    from typing import Iterator, Mapping, MutableMapping, Sequence
 
 if sys.version_info >= (3, 10):
     from typing import TypeGuard
@@ -231,10 +242,109 @@ DataLike = Union[
     Tuple["DataEntryLike", ...],
 ]
 
+
+T = TypeVar("T", np.int64, np.int32)
+
+
+class FaceList(Generic[T], Sequence["np.ndarray[tuple[int], np.dtype[T]]"]):
+    indices: np.ndarray[tuple[int], np.dtype[T]]
+    values: np.ndarray[tuple[int], np.dtype[T]]
+
+    @overload
+    def __init__(
+        self,
+        values: np.ndarray[tuple[int], np.dtype[T]] | Sequence[T | int],
+        indices: np.ndarray[tuple[int], np.dtype[T]] | Sequence[T | int],
+    ) -> None: ...
+
+    @overload
+    def __init__(
+        self,
+        values: np.ndarray[tuple[int, int], np.dtype[T]]
+        | Sequence[np.ndarray[tuple[int], np.dtype[T]] | Sequence[T | int]],
+    ) -> None: ...
+
+    def __init__(
+        self,
+        values: np.ndarray[tuple[int], np.dtype[T]]
+        | Sequence[T | int]
+        | np.ndarray[tuple[int, int], np.dtype[T]]
+        | Sequence[Sequence[T | int] | np.ndarray[tuple[int], np.dtype[T]]],
+        indices: np.ndarray[tuple[int], np.dtype[T]] | Sequence[T | int] | None = None,
+    ) -> None:
+        if indices is None:
+            dtype = values.dtype if isinstance(values, np.ndarray) else np.int64
+            self.indices = np.empty(len(values) + 1, dtype=dtype)
+            self.indices[0] = 0
+            self.indices[1:] = np.cumsum([len(v) for v in values])  # type: ignore [arg-type]
+            self.values = np.concatenate(values)  # type: ignore [arg-type, assignment]
+        else:
+            self.values = np.array(values)  # type: ignore [assignment]
+            self.indices = np.array(indices, dtype=self.values.dtype)  # type: ignore [assignment]
+
+    @overload
+    def __getitem__(self, index: int) -> np.ndarray[tuple[int], np.dtype[T]]: ...
+
+    @overload
+    def __getitem__(
+        self, index: slice
+    ) -> Sequence[np.ndarray[tuple[int], np.dtype[T]]]: ...
+
+    def __getitem__(
+        self, index: int | slice
+    ) -> (
+        np.ndarray[tuple[int], np.dtype[T]]
+        | Sequence[np.ndarray[tuple[int], np.dtype[T]]]
+    ):
+        if isinstance(index, slice):
+            msg = "Slicing FaceList is not yet implemented."
+            raise NotImplementedError(msg)
+
+        if index < 0:
+            index += len(self)
+        if index < 0 or index >= len(self):
+            msg = f"Index {index} out of range for FaceList of length {len(self)}"
+            raise IndexError(msg)
+        start = self.indices[index]
+        end = self.indices[index + 1]
[…]
