### 2025-05-15 #7633 @paugier
Yes, it is of course fine. Thanks @gerlero !
