### 2025-05-15 c532ecd
Bump version to 0.9.6
