### 2025-05-16 718c8bc
Improve internal loads parsing function
