### 2025-05-16 6ae8ba8
Fix type hints
