### 2025-05-16 a505dc4
Support parsing and serialization of face lists
