### 2025-05-16 #7633 @mbarzegary
@gerlero given the green light of the reviewers, we will now work towards processing this for acceptance in JOSS. So please:

* Check and merge my PR with minor edits on the paper
* Work on the author's points of the final checklist I created ☝️
