### 2025-05-16 21a498b
Bump version to 1.0.0
