+++ foamlib/__init__.py
@@ -1,6 +1,6 @@
 """A Python interface for interacting with OpenFOAM."""
 
-__version__ = "0.9.7"
+__version__ = "1.0.0"
 
 from ._cases import (
     AsyncFoamCase,
