### 2025-05-16 #7633 @gerlero
> Double check authors and affiliations (including ORCIDs)

Double checked!

> Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.

https://github.com/gerlero/foamlib/releases/tag/v1.0.0

> Archive the release on Zenodo/figshare/etc and post the DOI here.

https://doi.org/10.5281/zenodo.15442960

> Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.

Checked!

> Make sure that the license listed for the archive is the same as the software license.

Checked!

@mbarzegary Thanks again. All done!
