### 2025-05-19 408e811
Bump foamlib from 0.8.1 to 1.0.0
Bumps [foamlib](https://github.com/gerlero/foamlib) from 0.8.1 to 1.0.0.
- [Release notes](https://github.com/gerlero/foamlib/releases)
- [Commits](https://github.com/gerlero/foamlib/compare/v0.8.1...v1.0.0)

---
updated-dependencies:
- dependency-name: foamlib
  dependency-version: 1.0.0
  dependency-type: direct:production
  update-type: version-update:semver-major
...

Signed-off-by: dependabot[bot] <support@github.com>
