+++ benchmark/requirements.txt
@@ -1,2 +1,2 @@
-foamlib==0.8.1
+foamlib==1.0.0
 PyFoam==2023.7
